taxbot/pyproject.toml
[tool.poetry]
name = "taxbot"
version = "0.1.0"
description = "Bot for changing tax payment passwords"
authors = ["TJ Palanca <mail@tjpalanca.com>"]
license = "AGPL-3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
robotframework-browser = "^18.2.0"
rpaframework = "^28.4.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.22.2"
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
